requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-docx==1.1.2
reportlab==4.2.2
requests==2.32.3
